+++ dashboard/src/features/projects/overview/components/ProjectHealthBadge/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectHealthBadge } from './ProjectHealthBadge';
