+++ dashboard/src/features/projects/overview/components/RunServicesAccordion/index.ts
@@ -1,0 +1,1 @@
+export { default as RunServicesAccordion } from './RunServicesAccordion';
