+++ dashboard/src/features/projects/overview/components/ServiceAccordion/index.ts
@@ -1,0 +1,1 @@
+export { default as ServiceAccordion } from './ServiceAccordion';
